First impl of Operation logic (for Move and Arrival)
There are still many gaps, for instance Move should be able to
split Goods, but it's not implemented yet, and that will lead to
our first complications, and of course we should implement
Departure and Unpack/Pack very soon.

Had to make symbolic links of the tests under operation/ to test/
otherwise they won't run, which looks weird if one consider that
previously the tests in operation/tests/__init__.py was actually
executed, and a bit less weird if one thinks that for python it
was just importing 'wms_core.operation.tests' instead of
'operation.tests.test_operation'

For now this does not impediment progress too much, we'll see if that's
a bug in nose or in the anyblok plugin.