Move: avoid deletion of original Goods record
In case of full moves planned then executed,
we had a new Goods record at the destination and
deletion of the original one.
But some other Operation might hold a reference to the
original one, leading to a FK constraint error.
Therefore it's better to switch it to 'past'.

This is good enough for the time being, but it'll generate
lots of littering of the database with past Goods records, and
is not consistent with Moves done in one shot, nor with my idea
that a Move is a somewhat transparent operation for actual
('present') Goods.

See the long comment in execute_after_split() for suggested
improvements.