Oblivion for operations
obliviate() is another recursive method, actually quite
close to cancel(). In a later version, it'll probably
fall back to cancel() for downstream ops that are only planned.

Also this pointed out that we need a finer tracking of incoming
goods for operations, which is done in this commit for those that
inherite the MultipleGoods mixin only (but hints at possible
future refactors)