At last, introduced dates and times
On Goods, dt_from and dt_until define the interval during which
the goods are, were or will be present (almost true that it'd be
at a location, except moves can update location without creating
a 'past' record)

On Operation, dt_execution gives the date/time at which the operation
has been or is supposed to be executed

Location.quantity() takes a datetime if not restricted to the
'present' state, and actually requires it.
Therefore, we don't need negative quantities any more, stock level
computations now simply use the datetimes to give an exact assessment
(still can have the exception wrt precise location due to moves
created in state=done, but not for long).