Check constraints for quantities
Now that we don't rely onto negative counterparts at all anymore,
we can add the constraint.
Not sure it's that useful (will help to mitigate some bugs, probably)
In the meanwhile it proves that the job of converting to datetime range
queries is complete, since the tests pass and they have full coverage.