Introduced "outcomes" Python property for SingleGoodsOperations
It'll really help to clarify both the main impl and the tests.
The latter are a sign that user libs will also benefit from this change.
Finally it's spelled plural for uniformity because it'll probably climb
up its way in the mandatory subclass API.

Had to change name on split to lift ambiguity with the preexisting
'get_outcome()' method (made again a Python property for consistency)