Finally Moves don't mutate the Goods they're working on.
This means that any Move will now left once executed a 'past' Goods record,
while before that was only true for Moves first planned, then executed.

This may seem bold and an unnecessary volumetry in the database but

- it won't change much if most of the Moves get planned first
- it's the only way to get accurate results in Location.quantity()
  while peeking at past or future records. For some applications,
  it may make a difference
- more generally, it's much better for the conceptual consistency
- there will be ways to simplify past history
- with GiST indexes (for timestamp ranges), PostgreSQL should not
  fear a few million Goods (Avatar) records

Also updated the doc, and found a few unrelated miswordings, or needed
clarifications.