all Operations 'goods' field become the common 'working_on'
Now all Operations have the working_on m2m, with the caveat that
it uses, until we fix the 'enriched m2m' pattern, two tables
(one for the m2m, one for additional fields), but that is somewhat
burried within the base Operation class.

WmsSingleGoodsOperation is kept, mostly as a convenience,
WmsMultipleGoodsOperation is actually merged in the base class.

Many cases of goods_orig_dt_until cleaned up and also, quantity is
now available only on WmsSingleGoodsSplitterOperation.

What remains: cleanups, reorg of the doc, maybe cancel_single() and
obliviate_single() could get a base implementation (it's almost always
the same: restore dt_until and reason, and delete the outcomes).

Also, Move doesn't need origin any more (but that's due to the previous
refactor, that made it create 'past' Goods in all cases), so we also
removed that.