Default impls for cancel_single/obliviate_single
Since the realignment of created-done move, it's become
really uniform. Now that the base class takes care of saving
what is not part of the new Goods line creation, a common
implementation is enough.

We are very close to the Avatars refactor ; actually it should
become almost trivial (we'll see)

Still creative/destructive ops benefits from an even simpler one,
and that's one query less for poor DB.