Introduction of Avatars (no doc)
The Goods Model has been split in two according to
the plan still present in doc/improvements.rst

The issue with quantities is merely ignored. I can't
really decide if quantity should go to Avatar as well, so for
now it doesn't, and Split/Aggregate create new Goods records,
which feels very wrong. Maybe quantity should be both on Goods and
Avatars ? Would have to think of it a bit more, but I'd say
I'm ripe for getting rid of quantity in wms_core

As expected, the change isn't so drastic on the main code (see
the default impl for cancel_single() and obliviate_single(),
though), but much adaptation was needed in tests.
Many tests were rewritten using higher level methods, like
create() instead of insert(), and the 'outcomes' Python property
has been really useful