wms-reservation: locking and reserving
This is the second part of the reservation logic:
looking up Goods satisfying the Requests and actually
reserving them.
It comes with everything that should be needed reserver
processes, notably locking.

Except in the largest applications, reservations should be
done in-order, but there is some capability to cut the stack in
independent parts, by using the query_filter parameter in reserve_all().
Of course it's then the caller's responsibility that (functionally) the
ordering stays acceptable.