Starting the refactor of putting quantities aside.
At this point, the 'quantity' field is moved, as well as
Split/Aggregate ops and the Splitter mixin. The latter being now
lower in the class __mro__, we don't need the
execute_planned_after_split special method, and currently we
even have all Splitter ops not needing to be aware of the split
in their execute_planned (was different when move was an exception,
before avatars).

For now, all tests pass (except test_str for Goods) on a DB that
has all 3 bloks installed.

Next step : separation of operation tests that run with splits from
wms_core and rewrite of wms_core tests without quantities. At this
point, tests will have to be run with the proper modules installed.
Namely, wms-core's without wms-quantity