Location tags and recursive queries
Introducing the mono-valued tag field of Model.Wms.Location,
and a generic CTE to recurse across the hierarchy while applying
defaulting of missing tag onto the parent.

The next step would be to allow for quantity() to become recursive by
joining on the CTE and maybe moving this method to
Model.Wms.Goods.Avatar, making the location orientation just an option
among others.