Location.quantity(): recursion and tag filtering
Integrated the recursive CTE and the tag filtering capability
it provides.

Also fixed the CTE for the case where the start of recursion itself must
inherit its tag: it used to put None for the whole subhierarchy, and
even if it seems logical, it's too much inconsistent for callers.

Changed tag field to Text because that's what the literal in the CTE
would be.
Actually Anyblok Strings are always varchar, whereas this is useless in
most cases for PostgreSQL, hence most of our Strings should be changed to
Text.