Properties: as_dict() and tightened duplicate()
Before this commit, inner mutability of flexible properties after
a call to duplicate() would have mutated the original.

Note that if some field property happens to be a relationship, then
similar issues appear: if it is a Many2One, then mutating the target will do
it for both copies, but in that case I expect the caller to have done this
on purpose, since that's the point of relationships.