Properties dict API: brackets operator and get()
On a Properties instance, one can set and retrieve the value with
the brackets operator; now get() uses an optional positional argument
instead of a keyword one (as dicts do).

We can't readily subclass UserDict, because it uses the ABC metaclass,
and that clashes with SQLAlchemy's metaclass. There's probably a way to
cheat, but it's not obvious and would certainly introduce some frailness.