Unpack: direct properties and local ids
With this change, it's now possible to specify directly the
properties on each outcome of an Unpack (besides simply forwarding
them from the input).

Also, the special `local_goods_ids` parameter allows to choose the
underlying Goods of each outcome in advance. This is mostly useful
for chains of Pack/Unpack, and will probably be renamed in the
future.

Finally `Goods.update_properties()` is used where relevant, for
potential better efficiency.