Goods Type Properties
Related to the introduction of the hierarchy of Goods Types, this commit
now introduce Properties for Types, so that for instance Operations
working on parent Types can base Type-dependent logic on properties.

In short, now the Goods Properties default on the Type's, which
themselves are cascading upwards.

Since we expect Properties to be mostly scalar, there is no attempt
to merge those that would happen to be be dicts, but that may appear in
the future as well