Using SharedDataTestCase
WmsTestCaseWithGoods becomes a SharedDataTestCase, so that the goods
records, the arrival etc are only created once per test class.
This is in advance of Anyblok, to help mature it.

In each test module, we have to delete the imported base class at the
end, otherwise Nose sees it, transplant it in the surrounding module
(making a copy to share a potential setup_module) and executes its
setUpClass even though it bears no concrete test.