Assembly specification refactor: preparations
With this change, the matching logic is more separated from the
building of the outcome properties.

The new match_inputs() method does the matching, and does not
attempt to extract properties for forwarding: this is now done
by forward_properties()

We leverage the recently introduced 'match' field to avoid passing
the match results from one internal method to the other. This paves
the way for the wished match/check dichotomy.

Also, 'state' arguments are introduced, even though they aren't
currently used (waiting for final form of specification to implement
that)

Also, it seems that global Property requirements were not
really applied (no exception raised). Introduced the corresponding
exception and raised it.