Assembly specification refactor: outcome_properties
This replaces the old 'properties' and 'properties_at_execution' with
the per-state parameter.

We now have some code duplication between build_outcome_properties() (called
in after_insert()) and execute() but that should be solved in the next round
(per-input specs wihpotential rematching) in which I expect execute() to
actually call build_outcome_properties(), and therefore also potentially
updating the 'contents' property, which is for now only filled from creation.