Assembly specification refactor: per input prop requirements
This commit implement the per state structure of each
of the input specifications, as previously promised in
docstrings and the check/match logic.

The check/match choice is not specified as planned (that was
stupid to put it in each input spec), but rather in the new
`inputs_spec_type` toplevel parameter ('check' or 'match' for
each state). Also hoping it'll be more intuitive than the former
'requirements'.

What remains to be done is the properties forwarding logic
with the new per-state specification, including the special
case of the `contents` property.