Assembly spec refactor: per input properties forward
This completes the refactoring program to have parameters
for working on Properties depend on the state transition
been made.

The `contents` property is been rewritten at each transition.
Maybe something smarter can be done later, but that'd be
micro-optimisation anyway.

The `build_outcome_properties` method is renamed to
`outcome_properties` and now returns a dict, for Wms.Goods.Property
creation or update.