Goods: independent launching of tests with Nose plugin
test_goods is being moved inside goods.tests, and the import logic
is changed so that it's possible to launch these tests only with the
plugin (hence gaining proper coverage for this package).

This means that core.goods must be directly importable.
The new idiom is actually nicer (less # noqa tags), could be generalized later
to the operation package.