Made tests of Avatar str/repr timezone independent
and also fixed a missing brancket in repr that I'd never
noticed

In a previous simplifying change, the tested Avatar creation
had been delegated to WmsTestCaseWithGoods, and that seems to
change the way the roud-trip with the database behaves with
respect to timezones. Now the given date/time (self.dt_test1)
is most of the times reinterpreted in the server timezone, not changing the
absolute time, nor failing an equality test, but affecting the
display. That explains my previous hasty change of the tested
value, leading to a failure on CI.

All this being very erratic, it's best not to depend on the
timezone, and str/repr of the datetime object is not what we
are testing anyway.

To illustrate how funny it can be, something seen with a pdb.set_trace()
in Arrival.after_insert : while in after_insert(), self.outcomes[0].dt_from
is the original (hence in UTC, fixedoffset=0), it's been rewritten in
the server timezone after the return