Removal of Wms.Location: wms-core tests pass
The recursive quantity query has been adapted to go through avatars
and for the change of name of the 'tag' field. For now it works only
with 'present' Avatars, which means that future or past queries won't
take moves of locations (a.k.a containers) into account (this will
require much more testing)

The structure of Python modules has stayed the same, although it's clear
that some regrouping would be more consistent at this point.

wms-reservation and wms-quantity aren't adapted at all yet