Recursive quantity queries and location moves.
With this change, while peeking in the future or looking into the past,
the recursive quantity query takes changes of the location hierarchy
into account, i.e., the filtering by states and datetimes is done
at all levels of the recursion, and not only for the Goods been counted.

The resolution of the top level tag can still be false due to changes in
the hierarchy, because location tags could actually disappear quite
soon.

This is also a net benefit from been able to track location moves: with
the former Location model, it was possible to change the hierarchy, but
this was transversal to states and time information.

It's quite possible that both the methods involved and the query itself
could be simplified, but that's left to a later refactor.