New 'container' behaviour, used for now as marker
This behaviour is mandatory for all Goods that can contain other Goods,
i.e., be the 'location' field of some Avatars.
It could be used later or in downstream code to
further precise what kind of Goods records, quantities etc can be in a
given container type. For now its presence is only checked in Operations
involving containers (Move, Arrival, Apparition, Teleportation).