Quantities: pluggable filtering (container type etc)
While less powerful than the previous tags system (considered to be
redundant given that containers now have types and properties), the
filtering by container types is good enough for current filtering needs
(excluding a trash area, for instance).

A more general pluggable filtering system is actually introduced, of
which filtering by type of the direct container is just a special case.

But this pluggable filtering does not apply to the recursive part of the
quantity queries.