Default dt_execution for planned Operations
It's been a requirement from the start that dt_execution would be
mandatory upon creation of planned Operations, but the real technical
requirement is that timeranges of Avatars for a given PhysObj record
don't overlap.

At the beginning of development, we've been shy of having very short
lived Avatars (or even with 0 lifespan), but we had to do this anyway
with Splits.

Now we think that it's better to have these funny looking Avatars and to
refine them later than to force developers to make them themselves in
their planners, as it brings only confusion and discomfort.