Removed Avatar.reason in favour of the clearer 'outcome_of'
This doesn't include a migration step (yet ?)

outcome_of is strictly about the origin of the Avatar, whereas
reason was about any mutation. Actually with the Avatar concept,
that led to all Operations creating new Avatars, it has become
somewhat unneeded. If one wonders why some Avatar's state is
'past' for instance, it's easy enough to look for Operations
having it as input. The 'reason' field on the other hand was
a remnant of when we thought that Operations would mutate their
inputs (before Avatars were designed)