Removed HistoryInput.latest_previous_op
It had become a data duplication of Avatar.outcome_of. For the time
being it'll be simpler not to have to maintain consistency on this in
DAG manipulation primitives.

Of course replacing it by joins will impact performance, and we'll
consider reintroducing that duplication in an appropriate way (
for instance, with a targetted foreign key) in the future.