alter_destination: changing an Operation outcome location
This is the first primitive to change planification.

This commit introduces the means to flag Operations that
are responsible for their outcomes locations, and in which
field that location is stored.

The `alter_destination()` method builds on that to change it,
with a recursive propagation to the followers, thanks to
the `input_location_altered` notification method.