Mixin for Operations with a single outcome
The reason I haven't done this before is that some early
experiment with several mixins failed.

Besides being more convenient for everybody,
this should reduce the dangerous usage of outcomes[0]
(outcomes pseudo-field is not guaranteed to be a list or tuple).
Recently, I've been preferring next(iter(outcomes)) but that's
a hard sell