Inventory: computing 'app' and 'disp' Actions
Actions are 'app' (will lead to Apparition),
'disp' (will lead to Disparition)
and 'telep' (will lead to Teleportation)

We still don't issue 'telep' Actions and could do with
more beefy tests, but that's a start.

Also played with various scenarios for direct outer joins
but finally it's simpler to go through the already existing
quantity queries, and there's even a good chance this could
work with wms-quantity with only few modifications.