Inventory Actions can issue Operations
All produced Operations are linked to the relevant
Inventory, thanks to the newly introduced field on the mixin

To create Disparitions and Teleportations,
`Inventory.Action.apply()` relies on a common `choose_affected()`
method, that's not yet written. It's mocked in the tests to check
that, indeed, apply() does issue all three types of Inventory
Operations correctly.

This `choose_affected()` will be the place where trickier logic will
stand, such as avoiding to affect reserved physical objects, etc.
In particular, it'd be the main override target for dependent Bloks.