Using timezone information everywhere
Whereas the round-trip with the database magically converts
all naive datetimes to datetimes with timezone information
based on the server's locale, but we need to be consistent
before hand, if only because we'll have to compare the
datetimes we produce with some from the database before flush.

We're using plain UTC for the time being because it doesn't matter
much: user interfaces should explicitely convert to the timezone
of the user anyway. We could change to the server's locale later,
but nothing says that this is especially relevant, even for an
Operation been created manually by an user.