WIP using a btree_gist exclude rule to avoid overlapping timespans
Note that this introduces a few new test failures, it's hard to
know at this point if there are previous bugs or not. The one
in core.physobj isn't a bug, and all tests of core.physobj now pass.

As for coverage, maybe later we'll need to test with or without
the `btree_gist` extension. For now, it seems reasonible to test
only with it.