Adapted Operation.alter_execution to avoid overlaps
This is a bit unpleasant as it requires some flushing
(don't know how to force SQLA to send requests in the right ordering)

The two cases (sooner and later) really have to be treated separately,
and that's the opportunity to cut out a separate method, too. Otherwise
cyclomatic complexity check would complain already.