Operation regular creations should have terminal inputs
By "regular creations", we mean "except during alteration
of planification"

By terminal input, we mean that right before the new Operation
creation, they aren't inputs of any Operation, and this
coincides now with the fact that `dt_until` is None.

This is much simpler: we don't need to store the original
value of `dt_until` any more. Actually that's a remnant of
pre-Avatar days: at this time, `dt_until`, being stored on the
PhysObj record, had to be updated in place.

This is also more robust: actually forbidding the case prevents
unwanted forks in the Operation DAG. The change in preparations
of `test_cancel_recursion` is actually an example of a bug
having to be fixed by the new rule